flask-backend/requirements.txt
blinker==1.8.2
cffi==1.17.0
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.33
typing_extensions==4.12.2
Werkzeug==3.0.3
